# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Maz Delivery API",
        "description": "Maz Delivery project",
        "contact": {
            "email": "info@mazdelivery.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "http://www.mazdelivery.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/inventory": {
            "get": {
                "tags": [
                    "inventory"
                ],
                "summary": "Get All Inventory",
                "description": "Get All Inventory",
                "operationId": "inventoryLists",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "inventory"
                ],
                "summary": "Post New Inventory",
                "description": "Post New Inventory",
                "operationId": "inventoryCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/{id}": {
            "get": {
                "tags": [
                    "inventory"
                ],
                "summary": "Show an Inventory",
                "description": "Show an Inventory",
                "operationId": "inventoryDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "inventory"
                ],
                "summary": "Update an Inventory",
                "description": "Update an Inventory",
                "operationId": "inventoryUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "inventory"
                ],
                "summary": "Delete an Inventory",
                "description": "Delete an Inventory",
                "operationId": "inventoryDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InventoryCreateFormRequest": {
                "title": "Inventory Create Form Request Fields",
                "description": "Inventory Create request body data",
                "required": [
                    "first_name"
                ],
                "type": "object"
            },
            "InventoryUpdateFormRequest": {
                "title": "Inventory Update Form Request Fields",
                "description": "Inventory Update request body data",
                "required": [
                    "first_name"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}